feat: separate out proc_macro from inner functions
