+++ crates/witgen_macro_helper/README.md
@@ -1,0 +1,3 @@
+# witgen_macro
+
+Macros for witgen which is a library to help you generate wit definitions in a wit file for WebAssembly
